Changes related to PR comments
